refactor service instance validation
