+++ service/service_bindings.go
@@ -1,0 +1,6 @@
+package service
+
+func mergeServiceBindings() {
+	// TODO: this needs to be implemented, parse SERVICE_BINDING_ROOT/<services> here..
+	// TODO: move this method into /service/service.go->Init->/service/service_bindings.go ?
+}
